+++ example/peripheral/can/canfd/src/canfd_polled_loopback_mode_example.c
@@ -288,8 +288,8 @@ static void FFreeRTOSCanfdSendTask(void *pvParameters)
                 printf("canfd%d send failed.\n", instance_id);
             }
             count[instance_id]++;
-            xQueueSendToBack(xQueue, &instance_id, portMAX_DELAY);
             vTaskDelay(CAN_SEND_PERIOD);
+            xQueueSendToBack(xQueue, &instance_id, portMAX_DELAY);
         }
         if ( (send_frame[instance_id - 1].canid == send_max_id) )
         {